Avoid memory pollution by not using a temporary List when possible
In the normal case, all rails will be filled in a single sweep. We can use
the rails list we already have for this, instead of creating a temporary list each time.
This fix helps reduce gc activity. In the abnormal case it now uses a LinkedList
since it is more consistent when adding new rail elements.